refactor: moves game logic to 'useGameLogic' hook
